compute ng script content -> run
